# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_request.CreateEmailTemplate:
    properties:
      config:
        additionalProperties: true
        type: object
      name:
        type: string
    required:
    - config
    - name
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_request.UpdateEmailTemplate:
    properties:
      config:
        additionalProperties: true
        type: object
      name:
        type: string
    type: object
...
  github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_response.EmailTemplateDetail:
    properties:
      config:
        additionalProperties: true
        type: object
      created_at:
        type: string
      name:
        type: string
      template_id:
        type: integer
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_response.EmailTemplateItem:
    properties:
      config:
        additionalProperties: true
        type: object
      created_at:
        type: string
      name:
        type: string
      template_id:
        type: integer
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_user_dto_request.Credentials:
    properties:
      login_id:
        description: 사용자 이름
        type: string
      password:
        description: 비밀번호
        type: string
    required:
    - login_id
    - password
    type: object
host: localhost:8080
info:
  contact:
    email: test@example.com
    name: 홍정완
  description: 채용 프로세스 엔진 백엔드 API 문서
  title: Recruit Process Engine API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: 로그인 아이디와 비밀번호로 인증하고 세션을 생성한다.
      parameters:
      - description: 로그인 정보
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_user_dto_request.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: message와 user_id 반환
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      summary: 사용자 로그인
      tags:
      - Auth
...
  /email-templates:
    get:
      description: 저장된 모든 이메일 템플릿의 정보를 반환한다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_response.EmailTemplateItem'
            type: array
        "500":
          description: Internal Server Error
      summary: 이메일 템플릿 목록 조회
      tags:
      - EmailTemplates
    post:
      consumes:
      - application/json
      description: 새로운 이메일 템플릿을 생성한다.
      parameters:
      - description: 생성할 템플릿 정보
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_request.CreateEmailTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_response.EmailTemplateDetail'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 이메일 템플릿 생성
      tags:
      - EmailTemplates
  /email-templates/{template_id}:
    delete:
      description: 선택된 이메일 템플릿을 삭제한다.
      parameters:
      - description: 템플릿 식별자
        in: path
        name: template_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 이메일 템플릿 삭제
      tags:
      - EmailTemplates
    get:
      description: 선택된 이메일 템플릿의 상세 정보를 반환한다.
      parameters:
      - description: 템플릿 식별자
        in: path
        name: template_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_response.EmailTemplateDetail'
        "404":
          description: Not Found
      summary: 이메일 템플릿 상세 조회
      tags:
      - EmailTemplates
    put:
      consumes:
      - application/json
      description: 이메일 템플릿 설정을 업데이트한다.
      parameters:
      - description: 템플릿 식별자
        in: path
        name: template_id
        required: true
        type: integer
      - description: 업데이트할 템플릿 정보
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_request.UpdateEmailTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_template_dto_response.EmailTemplateDetail'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 이메일 템플릿 수정
      tags:
      - EmailTemplates
...
swagger: "2.0"